# backend/swagger.json
{"openapi":"3.0.0","paths":{"/v1/calendar/generate":{"post":{"description":"Creates a complete content calendar with posts and comments based on the provided campaign configuration.","operationId":"CalendarController_generateCalendar_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCalendarDto"}}}},"responses":{"201":{"description":"Calendar generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCalendarResponseDto"}}}},"400":{"description":"Invalid campaign configuration"}},"summary":"Generate a new content calendar","tags":["calendar"]}},"/v1/calendar/generate-demo":{"post":{"description":"Quick way to test the system using pre-configured SlideForge campaign data.","operationId":"CalendarController_generateDemoCalendar_v1","parameters":[],"requestBody":{"required":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateDemoCalendarDto"}}}},"responses":{"201":{"description":"Demo calendar generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCalendarResponseDto"}}}}},"summary":"Generate a demo calendar using SlideForge sample data","tags":["calendar"]}},"/v1/calendar/generate-next":{"post":{"description":"Continues from the last generated calendar, automatically calculating the next week start date.","operationId":"CalendarController_generateNextWeek_v1","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCalendarDto"}}}},"responses":{"201":{"description":"Next week calendar generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCalendarResponseDto"}}}}},"summary":"Generate calendar for the next week","tags":["calendar"]}},"/v1/calendar/generate-next-demo":{"post":{"description":"Generates the next week calendar using SlideForge sample data, continuing from the last calendar.","operationId":"CalendarController_generateNextWeekDemo_v1","parameters":[],"responses":{"201":{"description":"Next week demo calendar generated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateCalendarResponseDto"}}}}},"summary":"Generate next week demo calendar","tags":["calendar"]}}},"info":{"title":"RedditMastermind","description":"API of RedditMastermind","version":"1","contact":{}},"tags":[{"name":"app","description":""}],"servers":[{"url":"http://localhost:3000"}],"components":{"schemas":{"CompanyInfoDto":{"type":"object","properties":{"name":{"type":"string","description":"Company name","example":"SlideForge"},"website":{"type":"string","description":"Company website","example":"https://slideforge.ai"},"description":{"type":"string","description":"Company description","example":"AI-powered presentation tool"},"valuePropositions":{"description":"Key value propositions","example":["10x faster presentations","AI-powered design"],"type":"array","items":{"type":"string"}}},"required":["name","website","description","valuePropositions"]},"PersonaWritingStyleDto":{"type":"object","properties":{"sentenceLength":{"type":"string","enum":["short","medium","long"],"example":"medium"},"usesEmojis":{"type":"boolean","example":false},"formality":{"type":"string","enum":["informal","neutral","formal"],"example":"informal"},"typicalPhrases":{"example":["honestly","in my experience"],"type":"array","items":{"type":"string"}}},"required":["sentenceLength","usesEmojis","formality","typicalPhrases"]},"PersonaDto":{"type":"object","properties":{"id":{"type":"string","example":"riley_ops"},"username":{"type":"string","example":"riley_techops"},"bio":{"type":"string","example":"Operations manager at a mid-size tech company"},"writingStyle":{"$ref":"#/components/schemas/PersonaWritingStyleDto"},"expertise":{"example":["operations","productivity"],"type":"array","items":{"type":"string"}},"tone":{"type":"string","enum":["casual","professional","enthusiastic","skeptical","academic"],"example":"casual"}},"required":["id","username","bio","writingStyle","expertise","tone"]},"SubredditDto":{"type":"object","properties":{"name":{"type":"string","example":"r/PowerPoint"},"rules":{"type":"array","items":{"type":"string"}},"typicalTopics":{"example":["presentations","templates"],"type":"array","items":{"type":"string"}},"audienceType":{"type":"string","example":"Business professionals"}},"required":["name","typicalTopics","audienceType"]},"KeywordDto":{"type":"object","properties":{"id":{"type":"string","example":"K1"},"term":{"type":"string","example":"AI presentation maker"},"searchVolume":{"type":"number","example":1200},"priorityScore":{"type":"number","example":8,"minimum":1,"maximum":10}},"required":["id","term","priorityScore"]},"GenerateCalendarDto":{"type":"object","properties":{"companyInfo":{"$ref":"#/components/schemas/CompanyInfoDto"},"personas":{"type":"array","items":{"$ref":"#/components/schemas/PersonaDto"}},"subreddits":{"type":"array","items":{"$ref":"#/components/schemas/SubredditDto"}},"keywords":{"type":"array","items":{"$ref":"#/components/schemas/KeywordDto"}},"postsPerWeek":{"type":"number","example":3,"minimum":1,"maximum":10},"weekStartDate":{"type":"string","description":"ISO date string","example":"2024-01-15"}},"required":["companyInfo","personas","subreddits","keywords","postsPerWeek"]},"PostCandidateDto":{"type":"object","properties":{"id":{"type":"string","example":"post_abc123"},"title":{"type":"string","example":"Best AI Presentation Tools in 2024?"},"body":{"type":"string","example":"Looking for recommendations..."},"subreddit":{"type":"string","example":"r/PowerPoint"},"targetKeywords":{"example":["K1","K3"],"type":"array","items":{"type":"string"}},"opPersonaId":{"type":"string","example":"riley_ops"},"potentialImpact":{"type":"number","example":75},"generatedAt":{"type":"string","example":"2024-01-15T10:30:00.000Z"}},"required":["id","title","body","subreddit","targetKeywords","opPersonaId","potentialImpact","generatedAt"]},"CommentPlanDto":{"type":"object","properties":{"id":{"type":"string","example":"comment_xyz789"},"personaId":{"type":"string","example":"jordan_consults"},"text":{"type":"string","example":"I've tried a bunch of these..."},"replyTo":{"type":"string","example":"root","description":"root or comment ID"},"delayMinutes":{"type":"number","example":21},"mentionsProduct":{"type":"boolean","example":true},"subtletyScore":{"type":"number","example":0.7,"minimum":0,"maximum":1}},"required":["id","personaId","text","replyTo","delayMinutes","mentionsProduct","subtletyScore"]},"ThreadPlanDto":{"type":"object","properties":{"id":{"type":"string","example":"thread_def456"},"post":{"$ref":"#/components/schemas/PostCandidateDto"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentPlanDto"}},"conversationStyle":{"type":"string","enum":["debate","agreement","question-answer","experience-sharing"],"example":"question-answer"},"estimatedEngagement":{"type":"number","example":25}},"required":["id","post","comments","conversationStyle","estimatedEngagement"]},"TimeSlotDto":{"type":"object","properties":{"date":{"type":"string","example":"2024-01-15T00:00:00.000Z"},"dayOfWeek":{"type":"string","example":"Monday"},"preferredHour":{"type":"number","example":14,"minimum":0,"maximum":23}},"required":["date","dayOfWeek","preferredHour"]},"HardConstraintsDto":{"type":"object","properties":{"subredditFrequency":{"type":"boolean","example":true},"personaNoSelfReply":{"type":"boolean","example":true},"weeklyLimit":{"type":"boolean","example":true}},"required":["subredditFrequency","personaNoSelfReply","weeklyLimit"]},"SoftConstraintsDto":{"type":"object","properties":{"topicDiversity":{"type":"number","example":0.85,"minimum":0,"maximum":1},"personaDistribution":{"type":"number","example":0.9,"minimum":0,"maximum":1},"keywordCoverage":{"type":"number","example":0.75,"minimum":0,"maximum":1}},"required":["topicDiversity","personaDistribution","keywordCoverage"]},"ConstraintCheckResultDto":{"type":"object","properties":{"passed":{"type":"boolean","example":true},"hardConstraints":{"$ref":"#/components/schemas/HardConstraintsDto"},"softConstraints":{"$ref":"#/components/schemas/SoftConstraintsDto"},"warnings":{"example":[],"type":"array","items":{"type":"string"}}},"required":["passed","hardConstraints","softConstraints","warnings"]},"QualityCheckDto":{"type":"object","properties":{"name":{"type":"string","example":"ad_detection"},"passed":{"type":"boolean","example":true},"score":{"type":"number","example":0.85},"details":{"type":"string","example":"Content appears natural and not promotional"}},"required":["name","passed","score","details"]},"QualityReportDto":{"type":"object","properties":{"threadId":{"type":"string","example":"thread_abc123"},"overallScore":{"type":"number","example":82},"checks":{"type":"array","items":{"$ref":"#/components/schemas/QualityCheckDto"}},"passed":{"type":"boolean","example":true},"recommendations":{"example":["Consider varying sentence length"],"type":"array","items":{"type":"string"}}},"required":["threadId","overallScore","checks","passed","recommendations"]},"ScheduledThreadDto":{"type":"object","properties":{"id":{"type":"string","example":"entry_ghi012"},"thread":{"$ref":"#/components/schemas/ThreadPlanDto"},"slot":{"$ref":"#/components/schemas/TimeSlotDto"},"status":{"type":"string","enum":["draft","pending_review","approved","scheduled","posted","failed","rejected"],"example":"draft"},"constraintsSatisfied":{"$ref":"#/components/schemas/ConstraintCheckResultDto"},"qualityReport":{"$ref":"#/components/schemas/QualityReportDto"}},"required":["id","thread","slot","status","constraintsSatisfied"]},"CalendarMetadataDto":{"type":"object","properties":{"totalPosts":{"type":"number","example":3},"subredditDistribution":{"type":"object","example":{"r/PowerPoint":1,"r/SaaS":2},"description":"Posts per subreddit"},"personaUsage":{"type":"object","example":{"riley_ops":2,"jordan_consults":3},"description":"Comment count per persona"},"keywordsCovered":{"example":["K1","K3","K7"],"type":"array","items":{"type":"string"}},"qualityScore":{"type":"number","example":82},"diversityScore":{"type":"number","example":78}},"required":["totalPosts","subredditDistribution","personaUsage","keywordsCovered","qualityScore","diversityScore"]},"ContentCalendarDto":{"type":"object","properties":{"id":{"type":"string","example":"cal_jkl345"},"weekStartDate":{"type":"string","example":"2024-01-15T00:00:00.000Z"},"weekEndDate":{"type":"string","example":"2024-01-21T23:59:59.999Z"},"campaignId":{"type":"string","example":"slideforge_demo"},"entries":{"type":"array","items":{"$ref":"#/components/schemas/ScheduledThreadDto"}},"metadata":{"$ref":"#/components/schemas/CalendarMetadataDto"},"status":{"type":"string","enum":["generating","ready","approved","partially_posted","completed"],"example":"ready"},"createdAt":{"type":"string","example":"2024-01-15T10:30:00.000Z"},"updatedAt":{"type":"string","example":"2024-01-15T10:35:00.000Z"}},"required":["id","weekStartDate","weekEndDate","campaignId","entries","metadata","status","createdAt","updatedAt"]},"GenerateCalendarResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"calendar":{"$ref":"#/components/schemas/ContentCalendarDto"},"errors":{"example":[],"type":"array","items":{"type":"string"}},"warnings":{"example":[],"type":"array","items":{"type":"string"}},"generationTimeMs":{"type":"number","example":4523}},"required":["success","generationTimeMs"]},"GenerateDemoCalendarDto":{"type":"object","properties":{"weekStartDate":{"type":"string","description":"ISO date string for week start (defaults to current date)","example":"2024-01-15"}}}}}}